+++ lib/easy-rest/exceptions/api-error.d.ts
@@ -1,0 +1,7 @@
+export declare class HttpError extends Error {
+    protected status: number;
+    message: string;
+    constructor(status: number, message?: string);
+    getStatus(): number;
+    getMessage(): string | any;
+}