+++ lib/easy-rest/exceptions/bad-request-error.d.ts
@@ -1,0 +1,4 @@
+import { HttpError } from "./api-error";
+export declare class BadRequestError extends HttpError {
+    constructor(message?: string);
+}