+++ lib/easy-rest/exceptions/not-found-error.d.ts
@@ -1,0 +1,4 @@
+import { HttpError } from "./api-error";
+export declare class NotFoundError extends HttpError {
+    constructor(message?: string);
+}