+++ lib/easy-rest/exceptions/unauthorized-error.d.ts
@@ -1,0 +1,4 @@
+import { HttpError } from "./api-error";
+export declare class UnauthorizedError extends HttpError {
+    constructor(message?: string);
+}