+++ src/easy-rest/exceptions.ts
@@ -1,0 +1,4 @@
+export * from './exceptions/api-error';
+export * from './exceptions/bad-request-error';
+export * from './exceptions/not-found-error';
+export * from './exceptions/unauthorized-error';