+++ src/easy-rest/exceptions/bad-request-error.ts
@@ -1,0 +1,7 @@
+import {HttpError} from "./api-error";
+
+export class BadRequestError extends HttpError {
+  constructor(message?: string) {
+    super(400, message);
+  }
+}