+++ src/easy-rest/exceptions/not-found-error.ts
@@ -1,0 +1,7 @@
+import {HttpError} from "./api-error";
+
+export class NotFoundError extends HttpError {
+  constructor(message?: string) {
+    super(404, message);
+  }
+}