+++ src/easy-rest/exceptions/unauthorized-error.ts
@@ -1,0 +1,7 @@
+import {HttpError} from "./api-error";
+
+export class UnauthorizedError extends HttpError {
+  constructor(message?: string) {
+    super(401, message);
+  }
+}