+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Features
 * controller initialization via glob, no more need to list all controllers in ApplicationInstance constructor
+* added http exceptions support - see HttpError
 
 <a name="0.3.2"></a>
 # [0.3.2](https://github.com/Odrin/express-easy-rest/compare/0.3.1...0.3.2) (2016-08-31)